Added new logos, moved old images to a single folder and updated the new team image.
