+++ react-ui/src/login/Login.js
@@ -1,8 +1,8 @@
-import React from 'react';
 import axios from "axios";
+import React from 'react';
 import { Button, FloatingLabel, Form, Image } from "react-bootstrap";
 import { Navigate } from "react-router-dom";
-import Cleveland from "./Cleveland.png"
+import CleverLogo from "../images/CleverLogo.jpg";
 
 export default class Login extends React.Component {
     state = {
@@ -47,7 +47,7 @@ export default class Login extends React.Component {
         return (
             <div className="d-flex align-items-center justify-content-center vh-100 backgroundWhite">
                 <Form onSubmit={this.handleSubmit} style={{ width: "100%" }}>
-                    <Image src={ Cleveland } className='mb-3 col-sm-8 offset-sm-2'/><br/><br/>
+                    <Image src={ CleverLogo } className='mb-3 col-sm-8 offset-sm-2'/><br/><br/>
                     <h2 align="center">Login</h2><br/>
                     <FloatingLabel
                         controlId="formUsername"
